+++ c/CHANGELOG.rst
@@ -15,6 +15,10 @@ Refinements to the C API as we move towards 1.0.0. Changes:
 - Rename ``tsk_table_collection_reset_position`` to ``tsk_table_collection_truncate``
   and ``tsk_table_collection_record_position`` to ``tsk_table_collection_record_num_rows``.
 - Generalise ``tsk_table_collection_sort`` to take a bookmark as start argument.
+- Relax restriction that nodes in the ``samples`` argument to simplify must
+  currently be marked as samples. (https://github.com/tskit-dev/tskit/issues/72)
+- Allow ``tsk_table_collection_simplify`` to take a NULL samples argument to
+  specify "all samples in the current tables".
 
 ---------------------
 [0.99.0] - 2019-01-14